APIs/googleapis.com/dfareporting/v3.3/python-aiohttp-client/openapi_server/controllers/campaigns_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server.models.campaign import Campaign
from openapi_server.models.campaigns_list_response import CampaignsListResponse
from openapi_server import util


async def dfareporting_campaigns_get(request: web.Request, profile_id, id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None) -> web.Response:
    """dfareporting_campaigns_get

    Gets one campaign by ID.

    :param profile_id: User profile ID associated with this request.
    :type profile_id: str
    :param id: Campaign ID.
    :type id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str

    """
    return web.Response(status=200)


async def dfareporting_campaigns_insert(request: web.Request, profile_id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, body=None) -> web.Response:
    """dfareporting_campaigns_insert

    Inserts a new campaign.

    :param profile_id: User profile ID associated with this request.
    :type profile_id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param body: 
    :type body: dict | bytes

    """
    body = Campaign.from_dict(body)
    return web.Response(status=200)


async def dfareporting_campaigns_list(request: web.Request, profile_id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, advertiser_group_ids=None, advertiser_ids=None, archived=None, at_least_one_optimization_activity=None, excluded_ids=None, ids=None, max_results=None, overridden_event_tag_id=None, page_token=None, search_string=None, sort_field=None, sort_order=None, subaccount_id=None) -> web.Response:
    """dfareporting_campaigns_list

    Retrieves a list of campaigns, possibly filtered. This method supports paging.

    :param profile_id: User profile ID associated with this request.
    :type profile_id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param advertiser_group_ids: Select only campaigns whose advertisers belong to these advertiser groups.
    :type advertiser_group_ids: List[str]
    :param advertiser_ids: Select only campaigns that belong to these advertisers.
    :type advertiser_ids: List[str]
    :param archived: Select only archived campaigns. Don&#39;t set this field to select both archived and non-archived campaigns.
    :type archived: bool
    :param at_least_one_optimization_activity: Select only campaigns that have at least one optimization activity.
    :type at_least_one_optimization_activity: bool
    :param excluded_ids: Exclude campaigns with these IDs.
    :type excluded_ids: List[str]
    :param ids: Select only campaigns with these IDs.
    :type ids: List[str]
    :param max_results: Maximum number of results to return.
    :type max_results: int
    :param overridden_event_tag_id: Select only campaigns that have overridden this event tag ID.
    :type overridden_event_tag_id: str
    :param page_token: Value of the nextPageToken from the previous result page.
    :type page_token: str
    :param search_string: Allows searching for campaigns by name or ID. Wildcards (*) are allowed. For example, \&quot;campaign*2015\&quot; will return campaigns with names like \&quot;campaign June 2015\&quot;, \&quot;campaign April 2015\&quot;, or simply \&quot;campaign 2015\&quot;. Most of the searches also add wildcards implicitly at the start and the end of the search string. For example, a search string of \&quot;campaign\&quot; will match campaigns with name \&quot;my campaign\&quot;, \&quot;campaign 2015\&quot;, or simply \&quot;campaign\&quot;.
    :type search_string: str
    :param sort_field: Field by which to sort the list.
    :type sort_field: str
    :param sort_order: Order of sorted results.
    :type sort_order: str
    :param subaccount_id: Select only campaigns that belong to this subaccount.
    :type subaccount_id: str

    """
    return web.Response(status=200)


async def dfareporting_campaigns_patch(request: web.Request, profile_id, id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, body=None) -> web.Response:
    """dfareporting_campaigns_patch

    Updates an existing campaign. This method supports patch semantics.

    :param profile_id: User profile ID associated with this request.
    :type profile_id: str
    :param id: Campaign ID.
    :type id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param body: 
    :type body: dict | bytes

    """
    body = Campaign.from_dict(body)
    return web.Response(status=200)


async def dfareporting_campaigns_update(request: web.Request, profile_id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, body=None) -> web.Response:
    """dfareporting_campaigns_update

    Updates an existing campaign.

    :param profile_id: User profile ID associated with this request.
    :type profile_id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param body: 
    :type body: dict | bytes

    """
    body = Campaign.from_dict(body)
    return web.Response(status=200)
